% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solve.R
\name{item_assignment}
\alias{item_assignment}
\title{Solve an instance of MFC item assignment}
\usage{
item_assignment(distances, p, solver, positives, n_leaders_minority)
}
\arguments{
\item{p}{How many groups are to be created.}

\item{solver}{A string identifing the solver to be used ("Rglpk",
"gurobi")}

\item{positives}{logical - is item positively coded.}

\item{n_leaders_minority}{Scalar number -}

\item{items}{A data.frame of item features. Rows must correspond to
items and columns to features.}
}
\value{
A data.frame containing one column of item ids (here, the id
    corresponds to the order of the items) and one column contains the
    group assignments of the items. The original items are also
    returned as columns of the data.frame.
}
\description{
Solve an instance of MFC item assignment
}
