% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMFC.R
\name{matchMFC}
\alias{matchMFC}
\title{This is THE function}
\usage{
matchMFC(
  x,
  size,
  scales,
  positive_polarity = NULL,
  solver = "glpk",
  time_limit = 120,
  n_groups_both_polarity = NULL,
  similarity_matrix = FALSE
)
}
\arguments{
\item{x}{Matrix / data frame, the data input}

\item{size}{The size of the groups}

\item{scales}{Vector of length \code{nrow(x)}, scale affiliation of each item}

\item{positive_polarity}{Logical vector of length \code{nrow(x)}, is each item positively coded?}

\item{solver}{Optional. The solver used to obtain the optimal method.
Currently supports "glpk", "symphony", "lpSolve" and "gurobi".}

\item{time_limit}{Time limit in seconds, given to the solver. Default is 120 seconds. Set to 0 for no time limit.}

\item{n_groups_both_polarity}{Optional, number of groups that must have items with both polarities.}

\item{similarity_matrix}{Logical scalar, defaults to \code{FALSE}. Is \code{x} a similarity matrix (e.g., correlations).}
}
\description{
This is THE function
}
