% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILP.R
\name{item_assign_ilp}
\alias{item_assign_ilp}
\title{Construct the ILP represenation of an MFC item assignment problem}
\usage{
item_assign_ilp(
  distances,
  p,
  solver = "glpk",
  is_in_minority_class = NULL,
  n_leaders_minority = NULL
)
}
\arguments{
\item{distances}{An distance object or matrix representing the
distances between items}

\item{p}{The number of groups to be created}

\item{solver}{A string identifing the solver to be used ("Rglpk",
"gurobi")}

\item{is_in_minority_class}{Logical: Is item negatively coded?}

\item{n_leaders_minority}{numbers of negatively items that should be cluster leaders}
}
\value{
A list representing ILP formulation of the instance
}
\description{
Construct the ILP represenation of an MFC item assignment problem
}
\references{
T. Bulhões, G. F. de Sousa Filho, A. Subramanian, and F. C. Lucídio
dos Anjos, “Branch-and-cut approaches for p-cluster editing,”
Discrete Applied Mathematics, vol. 219, pp. 51–64, 2017.

T. Bulhões, A. Subramanian, G. F. Sousa Filho, and F. C. Lucídio dos
Anjos, “Branch-and-price for p-cluster editing,” Computational
Optimization and Applications, vol. 67, no. 2, pp. 293–316, 2017.

Grötschel, M., & Wakabayashi, Y. (1989). A cutting plane algorithm for a clustering problem.
Mathematical Programming, 45, 59-96.

Papenberg, M., & Klau, G. W. (2021). Using anticlustering to partition data sets into equivalent parts.
Psychological Methods, 26(2), 161--174. https://doi.org/10.1037/met0000301
}
